move up component logic
