+++ ui/admin/app/components/form/host-catalog/aws/index.hbs
@@ -1,1 +1,129 @@
-AWS Form
+<Rose::Form
+  @onSubmit={{@submit}}
+  @cancel={{@cancel}}
+  @disabled={{@model.isSaving}}
+  as |form|
+>
+  <form.input
+    @name='name'
+    @type='name'
+    @value={{@model.name}}
+    @label={{t 'form.name.label'}}
+    @helperText={{t 'form.name.help'}}
+    @error={{@model.errors.name}}
+    readonly={{false}}
+    as |field|
+  >
+    {{#if @model.errors.name}}
+      <field.errors as |errors|>
+        {{#each @model.errors.name as |error|}}
+          <errors.message>{{error.message}}</errors.message>
+        {{/each}}
+      </field.errors>
+    {{/if}}
+  </form.input>
+
+  <form.textarea
+    @name='description'
+    @type='description'
+    @value={{@model.description}}
+    @label={{t 'form.description.label'}}
+    @helperText={{t 'form.description.help'}}
+    @error={{@model.errors.description}}
+    disabled={{@model.isSaving}}
+    readonly={{false}}
+    as |field|
+  >
+    {{#if @model.errors.description}}
+      <field.errors as |errors|>
+        {{#each @model.errors.description as |error|}}
+          <errors.message>{{error.message}}</errors.message>
+        {{/each}}
+      </field.errors>
+    {{/if}}
+  </form.textarea>
+  <form.fieldset>
+    <:title>{{t 'form.type.label'}}</:title>
+    <:body>
+      <form.radioGroup
+        @name='aws-types'
+        @changed={{@toggleType}}
+        @selectedValue='dynamic'
+        as |radioGroup|
+      >
+        <radioGroup.card
+          @label={{t 'resources.host-catalog.types.static'}}
+          @value='static'
+          @helperText={{t 'resources.host-catalog.form.types.static.help'}}
+          @layout='wide'
+        />
+        <radioGroup.card
+          @label={{t 'resources.host-catalog.types.dynamic'}}
+          @value='dynamic'
+          @helperText={{t 'resources.host-catalog.form.types.dynamic.help'}}
+          @layout='wide'
+        />
+      </form.radioGroup>
+    </:body>
+  </form.fieldset>
+
+  {{#if @showDynamicPlugins}}
+    <form.fieldset>
+      <:title>{{t 'resources.host-catalog.form.provider-details'}}</:title>
+      <:body>
+        <form.radioGroup
+          @name='aws-plugins'
+          @selectedValue={{@model.compositeType}}
+          @changed={{route-action 'changeType'}}
+          as |radioGroup|
+        >
+          <radioGroup.card
+            @label='AWS'
+            @value='aws'
+            @helperText={{t 'resources.host-catalog.form.types.static.help'}}
+            @layout='tile'
+          />
+          <radioGroup.card
+            @label='Azure'
+            @value='azure'
+            @helperText={{t 'resources.host-catalog.form.types.dynamic.help'}}
+            @layout='tile'
+          />
+        </form.radioGroup>
+      </:body>
+    </form.fieldset>
+  {{/if}}
+
+  <form.input
+    @name='Region'
+    @type='region'
+    @value='region'
+    @label='Region'
+    readonly={{false}}
+  />
+
+  <form.input
+    @name='accesskey'
+    @type='accesskey'
+    @value='accesskey'
+    @label='Access key ID'
+    readonly={{false}}
+  />
+
+  <form.input
+    @name='secretaccesskey'
+    @type='secretaccesskey'
+    @value='secretaccesskey'
+    @label='Secret Access Key'
+    readonly={{false}}
+  />
+
+  {{#if (can 'save model' @model)}}
+    <form.actions
+      @disabled={{if @model.cannotSave @model.cannotSave}}
+      @enableEditText={{t 'actions.edit-form'}}
+      @submitText={{t 'actions.save'}}
+      @cancelText={{t 'actions.cancel'}}
+    />
+  {{/if}}
+</Rose::Form>
